Fix bugs and restructure directory hierarchy
Fixed a bug where directions would be in (-1, 0) instead of in (-1, +1),
causing some cars not to move in Freeway.

Restructured the directory hierarchy so that the public facing API only
has the goatar package, and not any sub-packages. This will prevent
users of GoAtar from using `game.Game`s as environments, when they
should really use the `Environment` struct.