Implement Breakout and return []float64 as state
Added the Breakout game.

Changed the implementation of `State()` to return a `[]float64` instead
of a `[]*mat.Dense`. The reason for this is that the state observation
will almost only ever be used in a neural network for prediction (or
some other prediction method, e.g. linear function approximation). The
previous implementation would required all matrices in the returned
slice from `State()` to be concatenated every time a prediction had to
be made. Because of this, we would have to allocate a new slice, and
then **copy** all the data from each matrix into the slice. This would
be super inefficient. Now, we just return a `[]float64` in row major
order of the state observation. Then, this slice can just be used as the
backing slice of a tensor or matrix, which will be a lot more efficient!

Added a new method `Channel()` which will return the matrix at some
argument channel index. This is useful if a GoNum matrix is needed for a
specific channel.