Add documentation and adjust bullet implementation
Added documentation to the helper methods, functions, and structs of
SeaQuest.

Changed the implementation of the bullet struct. Really, this struct had
no differences with the swimmer struct, so I've changed the
implementation to merge bullets with swimmers. The only difference is
that bullets in SeaQuest don't have a move timer, so I've left a method,
`newBullet` that returns a swimmer with the move timer initialized to 0.
In reality, some bullets may be faster than others, so the merging of
bullets and swimmers is okay - some bullets *could* have lower move
timers than others. `newBullet()` is simply a convenience method
for `newSwimmer(x, y, orientation, 0)`, since all bullets in this
implementation start being able to move, and never are unable to
move - all bullets travel the same speed.